Move `UpdateResponse` command
