+++ wcfsetup/install/files/lib/command/comment/response/UpdateResponse.class.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace wcf\system\comment\response\command;
+namespace wcf\command\comment\response;
 
 use wcf\data\comment\response\CommentResponse;
 use wcf\data\comment\response\CommentResponseAction;
@@ -15,7 +15,7 @@ use wcf\system\message\embedded\object\MessageEmbeddedObjectManager;
  * @author      Marcel Werk
  * @copyright   2001-2024 WoltLab GmbH
  * @license     GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
- * @since       6.1
+ * @since       6.2
  */
 final class UpdateResponse
 {
